Extract _git.py from _legacy.py
